% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{converter}
\alias{converter}
\title{Converter to read downloaded datasets and automatically put them into one large dataframe with xts}
\usage{
converter(file)
}
\arguments{
\item{file}{downloaded dataset}
}
\value{
list of annual/monthly/daily files
}
\description{
\code{converter} read/clean/write
}
