% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tables.R
\name{download_tables}
\alias{download_tables}
\title{Downloads tables}
\usage{
download_tables(output_file = "data.Rdata")
}
\arguments{
\item{output_file}{Name of the .Rdata file}
}
\description{
Downloads tables
}
