% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFdownload.R
\name{FFdownload}
\alias{FFdownload}
\title{Downloads Datasets from kenneth french's Website}
\usage{
FFdownload(output_file = "data.Rdata", tempdir = NULL,
  exclude_daily = FALSE, download = TRUE)
}
\arguments{
\item{output_file}{name of the .RData file to be saved (include path if necessary)}

\item{tempdir}{specify if you want to keep downloaded files somewhere save. Seems to be necessary for
reproducible research as the files on the website seems to change from time to time}

\item{exclude_daily}{excludes the daily datasets (are not downloaded) ==> speeds the rpocess up considerably}

\item{download}{set to TRUE if you actually want to download again. set to false and specify tempdir to keep processing the already downloaded files}
}
\value{
RData file
}
\description{
\code{FFdownload} returns an RData file with all datasets from Kenneth French's Website. Should help researchers
to work with the datasets and update the regularly. Allows for reprducible research.
}
\examples{
# FFdownload(output_file = "FF20180909.RData",tempdir = "C:/temp",exclude_daily = TRUE)
# load("FF20180909.RData")
# FFdownload$x_25_Portfolios_5x5$monthly$average_value_weighted_returns

}
